引入了全新的通信协议架构，并修改了核心的数据结构
## 1. High-Level Summary (TL;DR)

* **Impact:** **High** - 引入了全新的通信协议架构，并修改了核心的数据结构。

* **Key Changes:**

  * 将全局的 `task_id` 数据类型从 32 位 (`uint32_t`) 升级为 64 位 (`uint64_t`)，以支持更长的 UUID 标识。

  * 新增了基于状态机的帧解析协议层 (`protocol` 模块)，统一处理来自蓝牙和串口的命令。

  * 在 RTOS 任务管理器中新增了两个任务：`PROTO_CmdTask`（协议解析与命令分发）和 `PROTO_UartRxTask`（串口接收任务）。

  * 修改了蓝牙接收逻辑，将原本的直接处理函数 `BleApp_ReceivedUartStream` 替换为 `PROTO_OnRxBytes`，实现统一的协议入口。

+++ ZKFJ_BLE_wireless_uart_freertos/source/app/app_storage.h
@@ -24,7 +24,7 @@ extern "C" {
 // 任务元数据结构体
 typedef struct
 {
-    uint32_t task_id;                // 任务ID
+    uint64_t task_id;                // 任务ID
     uint32_t start_timestamp_ms;     // 开始时间戳（毫秒）
     uint32_t sample_period_ms;       // 采样周期（毫秒）
     uint32_t duration_ms;            // 任务持续时间（毫秒）
@@ -36,28 +36,28 @@ typedef struct
 // 初始化存储
 bool APP_Storage_Init(void);
 // 开始任务
-bool APP_Storage_BeginTask(uint32_t task_id, const app_storage_task_meta_t *meta);
+bool APP_Storage_BeginTask(uint64_t task_id, const app_storage_task_meta_t *meta);
 
 
 // 追加写入数据
-bool APP_Storage_AppendData(uint32_t task_id, const void *record, uint32_t record_size);
+bool APP_Storage_AppendData(uint64_t task_id, const void *record, uint32_t record_size);
 // 追加写入日志
-bool APP_Storage_AppendLog(uint32_t task_id, const void *data, uint32_t size);
-bool APP_Storage_LogPrintf(uint32_t task_id, const char *format, ...);
+bool APP_Storage_AppendLog(uint64_t task_id, const void *data, uint32_t size);
+bool APP_Storage_LogPrintf(uint64_t task_id, const char *format, ...);
 
 // 从日志文件偏移量读取数据
-int APP_Storage_ReadLog(uint32_t task_id, uint32_t offset, void *out, uint32_t size);
+int APP_Storage_ReadLog(uint64_t task_id, uint32_t offset, void *out, uint32_t size);
 // 获取日志大小
-bool APP_Storage_GetLogSize(uint32_t task_id, uint32_t *out_size);
+bool APP_Storage_GetLogSize(uint64_t task_id, uint32_t *out_size);
 
 // 从文件偏移量读取数据
-int APP_Storage_ReadData(uint32_t task_id, uint32_t offset, void *out, uint32_t size);
+int APP_Storage_ReadData(uint64_t task_id, uint32_t offset, void *out, uint32_t size);
 
 // 获取数据大小
-bool APP_Storage_GetDataSize(uint32_t task_id, uint32_t *out_size);
+bool APP_Storage_GetDataSize(uint64_t task_id, uint32_t *out_size);
 
 // 删除任务
-bool APP_Storage_DeleteTask(uint32_t task_id);
+bool APP_Storage_DeleteTask(uint64_t task_id);
 
 #ifdef __cplusplus
 }
